Add business_connection_id

+++ telebot_views/base.py
@@ -52,7 +52,7 @@ class Request:
         if self.callback:
             return self.callback.message
         if self.inline:
-            return Message(
+            message = Message(
                 self.__cached_user__.keyboard_id if self.__cached_user__ else -1,
                 from_user=self.inline.from_user,
                 date=int(now_utc().timestamp()),
@@ -61,6 +61,7 @@ class Request:
                 options={'text': self.inline.query},
                 json_string='',
             )
+            return message
         raise ValueError('Unknown type of request')
 
     async def get_user(self) -> UserModel:
@@ -127,7 +128,7 @@ class KeyboardMessageSender(EmptyMessageSender):
         if not text:
             return
 
-        if user.state.messages_to_delete:
+        if user.state.messages_to_delete and not self.view.request.message.business_connection_id:
             next_messages_to_delete = set(self.view.user_states.next_user_state.messages_to_delete)
             next_messages_to_delete -= set(user.state.messages_to_delete)
             await asyncio.gather(
@@ -146,6 +147,7 @@ class KeyboardMessageSender(EmptyMessageSender):
                     user.keyboard_id,
                     reply_markup=markup,
                     parse_mode=self.parse_mode.value or None,
+                    business_connection_id=message.business_connection_id,
                 )
             except ApiTelegramException as err:
                 if 'message to edit not found' in err.description:
@@ -153,8 +155,9 @@ class KeyboardMessageSender(EmptyMessageSender):
                     return await self.send()
                 raise err
         else:
-            if user.keyboard_id:
+            if user.keyboard_id and not self.view.request.message.business_connection_id:
                 try:
+                    assert user.keyboard_id is not None
                     await bot.bot.delete_message(message.chat.id, user.keyboard_id)
                 except ApiTelegramException as err:
                     if (
@@ -165,8 +168,8 @@ class KeyboardMessageSender(EmptyMessageSender):
                     else:
                         raise
 
-            keyboard = await bot.bot.send_message(
-                message.chat.id,
+            keyboard = await bot.bot_answer_message(
+                message,
                 text,
                 reply_markup=markup,
                 parse_mode=self.parse_mode.value or None,
@@ -350,7 +353,7 @@ class BaseView:
             await self.callbacks.answer_callback()
             await self.user_states.set()
 
-        if self.request.msg and self.delete_income_messages:
+        if self.request.msg and self.delete_income_messages and not self.request.message.business_connection_id:
             await bot.bot.delete_message(self.request.message.chat.id, self.request.message.message_id)
 
         return self.route_resolver.routes_registry[self.view_name]
